requirements.txt
langchain
langchain-google-genai
crewai
langchain-community
python-dotenv
crewai-tools

